several tests added, some fixes

+++ contracts/interfaces/IPlatformMinimal.sol
@@ -7,7 +7,7 @@ pragma solidity ^0.8.30;
 interface IPlatformMinimal {
 
     /// @notice функция возвращает базовый размер комиссии
-    function getBaseFee() external view returns (uint256);
+    function getBaseFee() external view returns (uint16);
     
     /// @notice Получить кампанию по глобальному индексу
     function getCampaignByIndex(uint32 index) external view returns (address);
